+++ src/sim/Tool_Info.cpp
@@ -20,7 +20,7 @@ void Tool_Info::drawTool() {
 		sim->getCursorMgr()->setCursor(SDL_SYSTEM_CURSOR_ARROW);
 	}
 	Coord pos = GetMousePosWorld();
-	int h = pos.getY();
+	float h = pos.getY();
 	if (currentTarget) {
 		h = currentTarget->drawInformation2D(pos.getX(), h);
 		//h = drawString(pos.getX(), h, "Nothing");
